+++ config/config.php
@@ -1,7 +1,9 @@
 <?php
 /**
- * Model User in Laravel App
+ * Model User in Laravel App, Address and Port for WebSocket Server
  */
 return [
     'user_model' => 'App\User',
+    'port'       => '8080',
+    'host'       => '0.0.0.0',
 ];