+++ src/Events/ConnectionClose.php
@@ -1,0 +1,23 @@
+<?php
+namespace YnievesDotNet\FourStream\Events;
+
+use App\Events\Event;
+use Hoa\Event\Bucket as Bucket;
+use Illuminate\Queue\SerializesModels;
+
+class ConnectionClose extends Event
+{
+    use SerializesModels;
+
+    public $message;
+
+    /**
+     * ConnectionClose constructor.
+     * @param Bucket $bucket
+     */
+    public function __construct(Bucket $bucket)
+    {
+        $this->bucket = $bucket;
+    }
+
+}